+++ src/renderer/entities/transaction/ui/TransactionTitle/TransactionTitle.test.tsx
