+++ src/renderer/pages/Basket/index.ts
@@ -1,0 +1,1 @@
+export { Basket } from './ui/Basket';
