+++ src/renderer/pages/Basket/ui/EmptyBasket.tsx
@@ -1,0 +1,13 @@
+import { Icon, BodyText } from '@shared/ui';
+import { useI18n } from '@app/providers';
+
+export const EmptyBasket = () => {
+  const { t } = useI18n();
+
+  return (
+    <div className="flex flex-col items-center justify-center gap-y-4 h-full">
+      <Icon as="img" name="emptyList" alt={t('basket.noOperationsLabel')} size={178} />
+      <BodyText className="text-text-tertiary w-[300px] text-center">{t('basket.noOperationsDescription')}</BodyText>
+    </div>
+  );
+};
