+++ src/renderer/shared/routes/paths.ts
@@ -30,6 +30,9 @@ export const Paths = {
 
   // Staking
   STAKING: '/staking',
+
+  // Basket
+  BASKET: '/basket',
 } as const;
 
 export type PathType = ObjectValues<typeof Paths>;
